requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LangChain
langchain==0.1.4
langchain-openai==0.0.5

# HTTP клиент
requests==2.31.0
httpx==0.26.0
beautifulsoup4==4.12.2

# Environment
python-dotenv==1.0.0

# Логирование
python-json-logger==2.0.7
